feat: create the functionality for decoding the last word number
